pyproject.toml
[tool.poetry]
name = "api-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
paho-mqtt = "^2.0.0"
pydantic = "^2.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
